core/views.py
from django.shortcuts import render
from django.http import JsonResponse
from rest_framework.decorators import api_view, permission_classes
from rest_framework.permissions import AllowAny
from rest_framework.response import Response


@api_view(['GET'])
@permission_classes([AllowAny])
def health_check(request):
    """
    Health check endpoint - no authentication required
    """
    return Response({
        'status': 'healthy',
        'message': 'TaskForce API is running',
    })
